Rename + Add to Register
