+++ services/vision/obstaclespointcloud/obstacles_pointcloud.go
@@ -1,6 +1,6 @@
-// Package radiusclustering uses the 3D radius clustering algorithm as defined in the
+// Package obstaclespointcloud uses the 3D radius clustering algorithm as defined in the
 // RDK vision/segmentation package as vision model.
-package radiusclustering
+package obstaclespointcloud
 
 import (
 	"context"
@@ -16,7 +16,7 @@ import (
 	"go.viam.com/rdk/vision/segmentation"
 )
 
-var model = resource.DefaultModelFamily.WithModel("radius_clustering_segmenter")
+var model = resource.DefaultModelFamily.WithModel("obstacles_pointcloud")
 
 func init() {
 	resource.RegisterService(vision.API, model, resource.Registration[vision.Service, *segmentation.RadiusClusteringConfig]{
@@ -29,26 +29,26 @@ func init() {
 			if err != nil {
 				return nil, err
 			}
-			return registerRCSegmenter(ctx, c.ResourceName(), attrs, actualR)
+			return registerOPSegmenter(ctx, c.ResourceName(), attrs, actualR)
 		},
 	})
 }
 
-// registerRCSegmenter creates a new 3D radius clustering segmenter from the config.
-func registerRCSegmenter(
+// registerOPSegmenter creates a new 3D radius clustering segmenter from the config.
+func registerOPSegmenter(
 	ctx context.Context,
 	name resource.Name,
 	conf *segmentation.RadiusClusteringConfig,
 	r robot.Robot,
 ) (vision.Service, error) {
-	_, span := trace.StartSpan(ctx, "service::vision::registerRadiusClustering")
+	_, span := trace.StartSpan(ctx, "service::vision::registerObstaclesPointcloud")
 	defer span.End()
 	if conf == nil {
-		return nil, errors.New("config for radius clustering segmenter cannot be nil")
+		return nil, errors.New("config for obstacles pointcloud segmenter cannot be nil")
 	}
 	err := conf.CheckValid()
 	if err != nil {
-		return nil, errors.Wrap(err, "radius clustering segmenter config error")
+		return nil, errors.Wrap(err, "obstacles pointcloud segmenter config error")
 	}
 	segmenter := segmentation.Segmenter(conf.RadiusClustering)
 	return vision.NewService(name, r, nil, nil, nil, segmenter)
